requirements.txt
flask==2.0.1
tensorflow==2.16.1
pillow==10.2.0
numpy>=1.26.0,<2.0.0
Werkzeug==2.0.2
pandas==2.2.0
openpyxl==3.1.2
